Removed some lines but didn't change any functionality, added the rest of the router tests, thing is now npm installable and, in fact installed on npm: https://www.npmjs.com/package/ruuter seems to work if you install it locally in a project and call 'var x = require('ruuter');'.
