+++ lib/writeFile.js
@@ -1,6 +1,4 @@
-var fs = require('fs');
-var mkdirp = require('mkdirp');
-var Buffer = require('buffer').Buffer;
+var fs = require('fs'), mkdirp = require('mkdirp'), Buffer = require('buffer').Buffer;
 function writeFile(path, ext, data, options,callback) {
   if ((path === undefined) || (ext === undefined) || (data === undefined) || (options === undefined)) {
     return 'Error:  All function parameters (path, extension, data, and options) except callback must be declared.';
@@ -18,10 +16,10 @@ function writeFile(path, ext, data, options,callback) {
   var fileName = path + name + ext;
   fs.readdir(path, function(err) {
     if(err) mkdirp.sync(path);
-    fs.access(fileName, function(errF) { // check file exists
-      if(errF) fs.writeFileSync(fileName, data);  // if file no exist, write file
-      if(options.overwrite) { // if user wants to overwrite
-        fs.access(fileName, fs.W_OK, function(errW) {  // check file is writeable
+    fs.access(fileName, function(errF) {
+      if(errF) fs.writeFileSync(fileName, data);
+      if(options.overwrite) {
+        fs.access(fileName, fs.W_OK, function(errW) {
           if(errW) throw errW;
           fs.writeFileSync(fileName, data);
           if (callback !== undefined) callback();
